Show bibliography button depends on existence bib
