+++ Angular/src/app/columns/columns.component.html
@@ -176,7 +176,10 @@
   [matMenuTriggerFor]="rightMenu"></div>
 
 <mat-menu #columnMenu="matMenu">
-  <button (click)="this.show_column_bibliography(this.current_column.documents)" mat-menu-item>
+  <button
+    *ngIf="this.current_column.has_bibliography"
+    (click)="this.show_column_bibliography(this.current_column)"
+    mat-menu-item>
     Column Bibliography
   </button>
   <button
