+++ Angular/src/app/services/bibliography-helper.service.ts
@@ -1,11 +1,44 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from '@oscc/api.service';
 
+import { Commentary } from '@oscc/models/Commentary';
+
 @Injectable({
   providedIn: 'root',
 })
 export class BibliographyHelperService {
   constructor(protected api: ApiService) {}
+
+  /**
+   * Retrieves all bib keys from the given commentary object
+   * @param commentary (Commentary)
+   * @return bib_keys (string[])
+   * @author Ycreak
+   */
+  public retrieve_bib_keys_from_commentary(commentary: Commentary): string[] {
+    const commentary_fields = [
+      'commentary',
+      'differences',
+      'apparatus',
+      'reconstruction',
+      'metrical_analysis',
+      'translation',
+    ];
+    let bib_keys: string[] = [];
+
+    // Retrieve all keys for the simple string commentary fields
+    for (const key in commentary) {
+      if (commentary_fields.includes(key)) {
+        bib_keys = bib_keys.concat(this.get_bib_keys(commentary[key]));
+      }
+    }
+    // Retrieve keys for the context fields
+    for (const i in commentary.context) {
+      bib_keys = bib_keys.concat(this.get_bib_keys(commentary.context[i].text));
+    }
+    bib_keys = [...new Set(bib_keys)];
+    return bib_keys;
+  }
 
   /**
    * Retrieves all bib_keys in the given string in a list
